17.05.2026 11:00

+++ src/main/java/com/example/Post/PostRepository.java
@@ -1,11 +1,10 @@
 package com.example.Post;
 
-import com.example.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface PostRepository
-        extends JpaRepository<Post,Integer> {
+        extends JpaRepository<Post,Long> {
 
 }
